add Checkout, Login Page, fix Cart Page

+++ src/pages/Cart.js
@@ -35,19 +35,21 @@ export default function Cart() {
   const cartItems = useSelector((state) => state.cart.items);
   console.log(cartItems);
   const totalPayment = useSelector((state) => state.cart.totalPayment);
-  console.log(totalPayment);
+  // console.log(totalPayment);
 
   const incrementHandle = (event) => {
     const existingItem = cartItems.find(
       (item) => item.id === event.target.parentNode.id
     );
     dispatch(cartActions.incrementQuantityFromCart(existingItem));
+    console.log(cartItems);
   };
   const decrementHandle = (event) => {
     const existingItem = cartItems.find(
       (item) => item.id === event.target.parentNode.id
     );
     dispatch(cartActions.decrementQuantityFromCart(existingItem));
+    console.log(cartItems);
   };
 
   return (
@@ -74,18 +76,27 @@ export default function Cart() {
                 // console.log();
                 return (
                   <tr className="text-center align-middle" key={item.id}>
-                    <td style={{ width: "10%" }}>
+                    <td style={{ width: "14%" }}>
                       <img src={item.img} alt={item.name} width="100%" />
                     </td>
-                    <td>{item.name}</td>
+                    <td className="fs-5 fw-bold">{item.name}</td>
                     <td>{formattedCurrency} VND</td>
                     <td>
                       <div className="w-100 d-flex align-item-center text-center">
-                        <Button variant="" className="w-25" id={id}>
+                        <Button
+                          variant=""
+                          className={`w-25 ${
+                            item.quantity === 0 ? "d-none" : "d-block"
+                          }`}
+                          id={id}
+                        >
                           <FontAwesomeIcon
                             icon={faPlay}
                             rotation={180}
                             id={id}
+                            className={`${
+                              item.quantity === 0 ? "d-none" : "d-block"
+                            }`}
                             onClick={decrementHandle}
                           />
                         </Button>
@@ -104,7 +115,11 @@ export default function Cart() {
                       <Button
                         variant=""
                         onClick={() => {
-                          dispatch(cartActions.removeItemFromCart(id));
+                          if (window.confirm("Are you sure ?")) {
+                            dispatch(cartActions.removeItemFromCart(id));
+                          } else {
+                            return;
+                          }
                         }}
                       >
                         <FontAwesomeIcon icon={faTrashCan} />
@@ -113,9 +128,16 @@ export default function Cart() {
                   </tr>
                 );
               })}
+              {cartItems.length === 0 && (
+                <tr>
+                  <td colspan="6" className="text-warning text-center">
+                    Cart is empty!
+                  </td>
+                </tr>
+              )}
             </tbody>
           </Table>
-          <Row className="justify-content-between">
+          <Row className="justify-content-between mt-5">
             <Col xs lg="6">
               <Button
                 variant=""
@@ -130,6 +152,7 @@ export default function Cart() {
             <Col xs lg="6" className="d-flex justify-content-end">
               <Button
                 variant="light"
+                className="border border-dark"
                 onClick={() => {
                   navigate("/checkout");
                 }}
@@ -142,15 +165,18 @@ export default function Cart() {
         </Col>
         <Col xs lg="4">
           <Card>
-            <Card.Body>
-              <Card.Title className="mb-2">CART TOTAL</Card.Title>
+            <Card.Body className="p-5">
+              <Card.Title className="mb-4">
+                <h4>CART TOTAL</h4>
+              </Card.Title>
               <Card.Text className="d-flex justify-content-between ">
                 <h5>SUBTOTAL</h5>
-                <p>{totalPayment} VND</p>
+                <p>{convertCurrency(totalPayment[0])} VND</p>
               </Card.Text>
+              <hr />
               <Card.Text className="d-flex justify-content-between">
                 <h5>TOTAL</h5>
-                <h5>19.779.000 VND</h5>
+                <h5>{convertCurrency(totalPayment[0])} VND</h5>
               </Card.Text>
               <Form.Control type="text" placeholder="Enter your coupon" />
               <Button variant="dark" className="text-center w-100 mt-2">
